+++ waterfall-cad/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to the
 
 ## Unreleased
 
+### Changed
+
+- The project now depends on (and supports) OpenCascade `7.8.0` or later
+
 ## 0.2.2.1
 
 ### Fixed